Refactor profile into service and components
