+++ types/funcionario.ts
@@ -20,3 +20,22 @@ export type RegistroDTO = {
   password: string;
   cargoId: number;
 };
+
+export type EstadoPerfilUsuario = "ACTIVO" | "INACTIVO";
+
+export interface UserProfile {
+  idFuncionario: number;
+  nombre: string;
+  apellidos: string;
+  cargo: string;
+  area: string;
+  email: string;
+  phone: string;
+  fechaIngreso: string;
+  ubicacion: string;
+  foto: string;
+  estado: EstadoPerfilUsuario;
+  fechaNacimiento: string;
+  oficina: string;
+  reportaA: string;
+}
